Changes on scripts
Added some bad physics for testing, change on the name of variables and functions. Rename for all the scripts. Added collision on the corners of the table to detect if the white ball has fall there. And finally try make the ball change direction when hit a wall (Works really bad).